+++ plugin.go
@@ -15,11 +15,44 @@ const (
 	PluginTypeProcessor = core.PluginTypeProcessor
 	PluginTypeSink      = core.PluginTypeSink
 	PluginTypeUtility   = core.PluginTypeUtility
+	PluginTypeConnector = core.PluginTypeConnector
 )
 
 type PluginInfo = core.PluginInfo
 type PluginCapabilities = core.PluginCapabilities
 type PluginConfig = core.PluginConfig
+type ConnectorClient = core.ConnectorClient
+type ConnectorResource = core.ConnectorResource
+type ConnectorRequest = core.ConnectorRequest
+type ConnectorRequestKind = core.ConnectorRequestKind
+type ConnectorResponse = core.ConnectorResponse
+type HealthPolicy = core.HealthPolicy
+type ProviderPolicy = core.ProviderPolicy
+type QuotaPolicy = core.QuotaPolicy
+type ResourceStatus = core.ResourceStatus
+type ResourceStatusEvent = core.ResourceStatusEvent
+type WarmupPolicy = core.WarmupPolicy
+
+const (
+	WarmupPolicyLazy     = core.WarmupPolicyLazy
+	WarmupPolicyPreload  = core.WarmupPolicyPreload
+	WarmupPolicyAlwaysOn = core.WarmupPolicyAlwaysOn
+
+	ResourceStatusUnknown      = core.ResourceStatusUnknown
+	ResourceStatusHealthy      = core.ResourceStatusHealthy
+	ResourceStatusDegraded     = core.ResourceStatusDegraded
+	ResourceStatusCongested    = core.ResourceStatusCongested
+	ResourceStatusDisconnected = core.ResourceStatusDisconnected
+	ResourceStatusReadonly     = core.ResourceStatusReadonly
+	ResourceStatusBlocked      = core.ResourceStatusBlocked
+	ResourceStatusRecovering   = core.ResourceStatusRecovering
+
+	ConnectorRequestRead       = core.ConnectorRequestRead
+	ConnectorRequestWrite      = core.ConnectorRequestWrite
+	ConnectorRequestBatchRead  = core.ConnectorRequestBatchRead
+	ConnectorRequestBatchWrite = core.ConnectorRequestBatchWrite
+	ConnectorRequestMetadata   = core.ConnectorRequestMetadata
+)
 
 type LogLevel = core.LogLevel